requirements.txt
fastapi
uvicorn
httpx
pydantic-settings
python-dotenv
slowapi
python-jose[cryptography]